move helpers to test file
